requirements.txt
anyio==3.6.2
astroid==2.14.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dill==0.3.6
fastapi==0.89.1
fastapi-simple-security==1.2
h11==0.14.0
httptools==0.5.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
pep8==1.7.1
platformdirs==2.6.2
pydantic==1.10.4
pylint==2.16.1
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.14.1
